Fixes bug that causes Windows "File / Save Settings" & the -od command line switch to form a malformed save file.
The Windows "File / Save Settings" command (which is the -od switch on the command line) has a bug in version 7.20, in that a malformed file will be written. Upon starting the program (or manually opening the saved file) the error message "Too few options to switch /F (1 given, 3 required)" will appear. That will cause the rest of the settings file after the error point to not be parsed or read in.

The bad line looks like “-YkO 84 84 Sta; Fixed stars” and is at or near line #146 in the file. There should be at least one space between "Sta" and ";", which will allow the semicolon comment marker to be seen as such, and will avoid the text "Fixed stars" being seen as an malformed instance of the -F command switch. A quick workaround is to simply delete that line, or manually insert a space. However, those steps will need to be done each time "Save Settings" is run.